user context initialisialized with register form datas set in local storage
