+++ frontend/src/context/RegistrationFormContext.jsx
@@ -1,52 +1,52 @@
-import React, { createContext, useContext, useMemo, useState } from "react";
-import PropTypes from "prop-types";
+// import React, { createContext, useContext, useMemo, useState } from "react";
+// import PropTypes from "prop-types";
 
-const FormContext = createContext();
+// const FormContext = createContext();
 
-export default function FormContextProvider({ children }) {
-  const [formData, setFormData] = useState({
-    pseudo: "",
-    email: "",
-    postal: "",
-    city: "",
-    password: "",
-    confirmation: "",
-  });
+// export default function FormContextProvider({ children }) {
+//   const [formData, setFormData] = useState({
+//     pseudo: "",
+//     email: "",
+//     postal: "",
+//     city: "",
+//     password: "",
+//     confirmation: "",
+//   });
 
-  const handleChange = (event) => {
-    const { name, value } = event.target;
-    setFormData((prevFormData) => ({
-      ...prevFormData,
-      [name]: value,
-    }));
-  };
+//   const handleChange = (event) => {
+//     const { name, value } = event.target;
+//     setFormData((prevFormData) => ({
+//       ...prevFormData,
+//       [name]: value,
+//     }));
+//   };
 
-  const updateUser = (field, value) => {
-    setFormData({ ...formData, [field]: value });
-  };
+//   const updateUser = (field, value) => {
+//     setFormData({ ...formData, [field]: value });
+//   };
 
-  const saveUserToLocalStorage = () => {
-    localStorage.setItem("user", JSON.stringify(formData));
-  };
+//   const saveUserToLocalStorage = () => {
+//     localStorage.setItem("user", JSON.stringify(formData));
+//   };
 
-  const contextValue = useMemo(
-    () => ({
-      formData,
-      setFormData,
-      handleChange,
-      updateUser,
-      saveUserToLocalStorage,
-    }),
-    [formData, setFormData]
-  );
+//   const contextValue = useMemo(
+//     () => ({
+//       formData,
+//       setFormData,
+//       handleChange,
+//       updateUser,
+//       saveUserToLocalStorage,
+//     }),
+//     [formData, setFormData]
+//   );
 
-  return (
-    <FormContext.Provider value={contextValue}>{children}</FormContext.Provider>
-  );
-}
+//   return (
+//     <FormContext.Provider value={contextValue}>{children}</FormContext.Provider>
+//   );
+// }
 
-FormContextProvider.propTypes = {
-  children: PropTypes.node.isRequired,
-};
+// FormContextProvider.propTypes = {
+//   children: PropTypes.node.isRequired,
+// };
 
-export const useFormContext = () => useContext(FormContext);
+// export const useFormContext = () => useContext(FormContext);
